+++ cmd/image/main.go
@@ -10,19 +10,19 @@ import (
 const (
 	ASPECT_RATIO = 4.0 / 3
 	FOV          = 50.0
-	RESOLUTION   = 1080
+	RESOLUTION   = 200
 )
 
 func main() {
-	camera := NewCamera(ASPECT_RATIO, FOV, CameraTransformation{
-		LookFrom: NewVector3(0, 5, -12),
-		LookAt:   NewVector3(0, 5, 5),
-		Up:       NewVector3(0, 1, 0),
-	})
+	//scene, camera := demo.CornellBox(ASPECT_RATIO, FOV)
+	scene, camera := demo.Bunny(ASPECT_RATIO, FOV)
 
-	scene := demo.CornellBox()
-	bvh := scene.Compile()
+	bvh := scene.CompileLBVH()
 	renderer := NewDefaultRenderer(bvh, camera)
+	renderer.Closest = UnlitClosestHitShader
+	renderer.Miss = SkyMissShader
+	renderer.Spp = 200
+
 	buff := NewBufferAspect(RESOLUTION, ASPECT_RATIO)
 	renderer.RenderToBuffer(buff)
 	//renderer.IntersectionHeatMapToBuffer(buff)