+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Implemented PHR BVH builder
+- benchmark package including all benchmarks conducted
+- Added optional unlit renderer
 ### Changed 
 - Fixed bug in enclosing functions
+- Parallelized PHR find initial Cut
 
 ## [0.0.3] - 2021-08-12
 ### Added